+++ pkg/operator/manager.go
@@ -29,11 +29,28 @@ import (
 	"knative.dev/pkg/logging"
 	controllerruntime "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/healthz"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/aws/karpenter/pkg/utils/injection"
 	"github.com/aws/karpenter/pkg/utils/options"
 )
+
+// Controller is an interface implemented by Karpenter custom resources.
+type Controller interface {
+	// Reconcile hands a hydrated kubernetes resource to the controller for
+	// reconciliation. Any changes made to the resource's status are persisted
+	// after Reconcile returns, even if it returns an error.
+	Reconcile(context.Context, reconcile.Request) (reconcile.Result, error)
+	// Register will register the controller with the manager
+	Register(context.Context, manager.Manager) error
+}
+
+// HealthCheck is an interface for a controller that exposes a LivenessProbe
+type HealthCheck interface {
+	LivenessProbe(req *http.Request) error
+}
 
 // NewManagerOrDie instantiates a controller manager or panics
 func NewManagerOrDie(ctx context.Context, config *rest.Config, opts *options.Options) manager.Manager {
@@ -60,6 +77,26 @@ func NewManagerOrDie(ctx context.Context, config *rest.Config, opts *options.Opt
 	}
 
 	return newManager
+}
+
+// RegisterControllers registers a set of controllers to the controller manager
+func RegisterControllers(ctx context.Context, m manager.Manager, controllers ...Controller) manager.Manager {
+	for _, c := range controllers {
+		if err := c.Register(ctx, m); err != nil {
+			panic(err)
+		}
+		// if the controller implements a liveness check, connect it
+		if lp, ok := c.(HealthCheck); ok {
+			utilruntime.Must(m.AddHealthzCheck(fmt.Sprintf("%T", c), lp.LivenessProbe))
+		}
+	}
+	if err := m.AddHealthzCheck("healthz", healthz.Ping); err != nil {
+		panic(fmt.Sprintf("Failed to add health probe, %s", err))
+	}
+	if err := m.AddReadyzCheck("readyz", healthz.Ping); err != nil {
+		panic(fmt.Sprintf("Failed to add ready probe, %s", err))
+	}
+	return m
 }
 
 func newRunnableContext(config *rest.Config, options *options.Options, logger *zap.SugaredLogger) func() context.Context {
